Big cleanup. Almost fresh start.
Most of the general funcitonality, MSL, chocie generation, welfare
calculation, has been moved to independent packages. This makes it
easier to distribute, takes less time to run as it doesn't need to be
recompiled everytime, and it makes the code of interest much cleaner.
Parsing stuff has been relegated to the backend where it belongs.